pyproject.toml
[project]
name = "gcp-ai-app-demo"
version = "0.1.0"
description = "GCP AI Dog Breed Detector - Identify dog breeds using Gemini AI"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.1",
    "firebase-admin>=7.2.0",
    "google-cloud-firestore>=2.23.0",
    "google-cloud-storage>=3.9.0",
    "google-genai>=1.65.0",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.41.0",
]
